+++ src/mruby_integration/models/gesture.cpp
@@ -2,6 +2,8 @@
 #include "mruby/class.h"
 #include "raylib.h"
 
+#include "mruby_integration/models/vector2.hpp"
+
 #include "ruby/models/gesture.hpp"
 
 struct RClass* Gesture_class;
@@ -38,6 +40,15 @@ mrb_Gesture_duration(mrb_state* mrb, mrb_value) -> mrb_value
   return mrb_float_value(mrb, GetGestureHoldDuration());
 }
 
+auto
+mrb_Gesture_dragged(mrb_state* mrb, mrb_value) -> mrb_value
+{
+  auto* return_vector = static_cast<Vector2*>(malloc(sizeof(Vector2)));
+  *return_vector = GetGestureDragVector();
+
+  return mrb_Vector2_value(mrb, return_vector);
+}
+
 void
 append_models_Gesture(mrb_state* mrb)
 {
@@ -55,6 +66,8 @@ append_models_Gesture(mrb_state* mrb)
                           MRB_ARGS_REQ(1));
   mrb_define_class_method(
     mrb, Gesture_class, "duration", mrb_Gesture_duration, MRB_ARGS_NONE());
+  mrb_define_class_method(
+    mrb, Gesture_class, "dragged", mrb_Gesture_dragged, MRB_ARGS_NONE());
 
   load_ruby_models_gesture(mrb);
 }